poc-llamaindex-sql/requirements.txt
llama-index>=0.11.0
llama-index-llms-gemini>=0.3.0
llama-index-embeddings-gemini>=0.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=7.4.0
